Refactor: 엔티티, 레포지토리 변경
