--- src/main/java/com/app/interstory/episode/domain/Collection.java
@@ -1,25 +1,0 @@
-package com.app.interstory.episode.domain;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@Table(name = "collection")
-public class Collection {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "id")
-	private Long id;
-
-	@Column(name = "user_id", nullable = false)
-	private Long userId;
-
-	@Column(name = "episode_id", nullable = false)
-	private Long episodeId;
-}
