+++ src/main/java/com/app/interstory/episode/repository/CartItemRepository.java
@@ -2,8 +2,8 @@ package com.app.interstory.episode.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.app.interstory.novel.domain.entity.Collection;
+import com.app.interstory.user.domain.entity.CartItem;
 
-public interface CollectionRepository extends JpaRepository<Collection, Integer> {
+public interface CartItemRepository extends JpaRepository<CartItem, Integer> {
 	boolean existsByUserAndEpisode(Long userId, Long episodeId);
 }
