--- src/main/java/com/app/interstory/episode/repository/EpisodeRepository.java
@@ -1,8 +1,0 @@
-package com.app.interstory.episode.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.app.interstory.episode.domain.Episode;
-
-public interface EpisodeRepository extends JpaRepository<Episode, Long> {
-}
