--- src/main/java/com/app/interstory/episode/repository/PaymentRepository.java
@@ -1,11 +1,0 @@
-package com.app.interstory.episode.repository;
-
-import java.util.List;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.app.interstory.episode.domain.Payment;
-
-public interface PaymentRepository extends JpaRepository<Payment, Long> {
-	List<Payment> findByUserId(Long userId);
-}
