+++ src/main/java/com/app/interstory/episode/repository/PointRepository.java
@@ -4,9 +4,9 @@ import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.app.interstory.episode.domain.Point;
+import com.app.interstory.user.domain.entity.Point;
 
 public interface PointRepository extends JpaRepository<Point, Long> {
-	List<Point> findByUserId(Long userId);
+	List<Point> findByUser(Long userId);
 }
