--- src/main/java/com/app/interstory/episode/repository/UserRepository.java
@@ -1,8 +1,0 @@
-package com.app.interstory.episode.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.app.interstory.episode.domain.User;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-}
